Improved header row detection
Previously it checks the length of the first column in the first and second row,
only if the lengths are different we consider the first row as header. However
this doesn't work for some Westpac accounts.

The improvement introduced in this commit is trying to identify the date column.
If there is no header row then both the first and second row should have a column
that stores dates, and they should have the same number of digits.